Add proper versioning for both the raw and the preprocessed profile format.
This setup allows us to make changes to either profile format and still keep
compatibility with old profiles.

When the raw profile format changes (due to a change in Gecko):
 - The version number (profile.meta.version) needs to be incremented in Gecko.
 - In raw-profile-versioning.js, CURRENT_VERSION needs to be set to that
   incremented version, and a conversion function from the old to the new
   version needs to be added to _updaters.
 - Profile preprocessing may need to be adjusted to parse the new raw
   profile format version.

When the preprocessed profile format changes (e.g. because a different data
format seems adequate, or because new data from the raw profile needs to be
accommodated):
 - preprocessProfile needs to be changed to output the new format.
 - In preprocessed-profile-versioning.js, CURRENT_VERSION needs to be
   incremented and an update function needs to be added to _updaters.

At all times, preprocessProfile only has code that converts the latest version
of the raw profile format into the latest version of the preprocessed profile
format. All compatibility code is hosted outside of preprocess-profile.js, in
the respective *-profile-versioning.js file.