Preprocessed format update for debugPath attribute
In bug 1329111 I decided to also add a debugPath attribute. But we've already
updated the version of the preprocessed profile format to a version that does
not have that field, and such profiles are already in the profile store, so
we need to do another update for adding this one field.

We don't need to increment the version of the raw profile format because it
hasn't shipped yet, and there exst no raw profiles of the hypothetical
in-between version that has debugName but not debugPath.