Fix stringTableIndexToNewFuncIndex cache.
If the raw profile contains already-symbolicated C++ frames (which is only
the case if you used a script to add symbols, like Talos profiling), then
we mutate funcNameIndex to point to just the function name. We then used
that new value of funcNameIndex to cache the result. This would cause a
subsequent cache lookup for the same function to fail.