Make thread ordering and hiding more consistent and saved to the URL
This commit changes the thread ordering state. Before a single
array for both the order and the hiding was created. Now these
are stored in two separate arrays for more consistent handling. In
addition the state was moved to the URL. The UX was improved and made
consistent throughout the various components, as well as guarding for
invalid application states like selecting a hidden thread, and hiding
all threads.