Call tree filters to transform stacks, and upgrade URLS
This commit changes call tree filters to use the idea of a transform
stack. The transforms in the stack will progressively modify the
thread's samples and stacks, filtering the profile according to a
variety of user-defined operations.

In addition this commit changes the URL structure. Rather than adding
another add-hoc if statement, it creates a versioning and upgrade
process to handle URL format changes.