Maintain the viewport vertical position in the flame graph
When doing operations such as changing the time selection or applying
a transform in the flame graph, it is a surprising user experience to
instantly scoll the viewport down to the bottom every time. This fixes
this by holding on to the viewport position even as these operations
are performed.