Parse files as JSON-encoded profiles by default
Before this patch and since #1025 we were parsing profiles as JSON only
if the received mime type was 'application/json'. To do this we relied
on the system to give us this mime-type. However this happens only if
the file name actually ends with '.json', and this isn't always the case
in our use cases.
So in this patch we revert to the old behavior where we parse as JSON if
we don't recognize a zip or gzip mime-type. We still require a valid
mime-type for these 2 other types though.

Fixes #1047