Fix WithSize calculating the size too early
The previous code used a ref callback function to start the process
updating the size after a resize. But a ref callback is synchronous and
so it happens before the full virtual DOM tree is rendered to the real
DOM, which is too early to get good measurements in all cases.

This commit changes this by using componentDidMount plus a
requestAnimationFame callback.