Render to canvas during React render and not via rAF.
If this results in too many renders, the rAF should be introduced around
the action dispatch, not around the render. That way it'll benefit both
canvas rendering and React DOM tree updates.

I'm removing this rAF because it made viewport zooming too sluggish. We
already wrap the action dispatch for zooming into rAF, and having another
rAF here meant another frame of latency.