Define a total order of call nodes.
Use this total order to adjust the y position of the selection highlighting
in the selected thread's activity graph. Instead of always having the
selected highlighting at the bottom of the category fill, have it somewhere
in the middle, based on the position of the selected call node in the call
tree.
This makes it so that, as you select different nodes in the call tree, unrelated
nodes will cause disjoint areas of the graph to highlight. And you can cover
the entire height of the fill an you go through the different call nodes.

Before this change, selecting sibling call nodes would sometimes highlight
the same parts of the graph, which could be a bit confusing.