Speed up execution of getFlameGraphTiming by avoiding allocations
Instead of calling CallTree.preloadChildrenCache, which allocates new
arrays for every parent in the call tree, implement a modified version
which stores everything needed in a couple of big arrays. Since the
new version is very specific to the flame graph, it is moved into
flame-graph.js instead, and CallTree.preloadChildrenCache is removed.