(do not review, see PR #1154) Upgrade webpack to version 4
This upgrade needed to happen due to a runtime error being generated by
the ModuleConcatenationPlugin where it would attempt to reassign a const
variable.

The largest change for this upgrade is that the Uglify plugin is no
longer encouraged, and it is preconfigured under the minimize option.
Unfortunately this meant that we lost the yarn start-prod-readable
command, as the configuration ignored attempts at not mangling variable
names.