Always reserve space for a vertical scrollbar.
Fixes #619, fixes #1123.

Depending on the number of threads / tracks visible in the list at the
top, and the height of the header area, the list of tracks can be
vertically scrollable. The scrollbar for vertical scrolling causes a
bunch of problems, depending on the platform:
 - Dragging the scrollbar thumb is currently impossible because we
   override the mouse events to control the range selection.
 - On platforms with overlay scrollbars, the scrollbar overlaps the
   graphs and makes it hard to interact with the parts of the graphs
   close to the right edge.
 - On platforms with classic (non-overlay) scrollbars, the scrollbar
   subtracts available area from the contents and causes the time scale
   of the graphs to be misaligned with respect to the ruler above the
   scroll box. Furthermore, changing a track's visibility can flip cause
   the track list to flip between having a scrollbar and not having a
   scrollbar, which makes the graphs shift.

This commit addresses all these problems by reserving a fixed amount of
space for the scrollbar at all times. It sets an absolute width on both
the timeline ruler at the top *and* on the scrollable contents.
When there is a scrollbar, the scrollbar is displayed in the reserved
space, otherwise the space is empty.
The reserved space is always 15px wide, regardless of the actual width
of the scrollbar on the current platform. If the actual scrollbar is
narrower than the reserved space, there will be a small gap. If it is
wider than the reserved space, it will cover some of the contents but
but will not affect the contents' size.

We use calc() and viewport units (100vw) in order to compute an absolute
CSS length value that we can use for the graph width. This is needed so
that we can give the "insides" of the scrollable element a width that is
the same regardless of the actual width of the scrollbar on the current
platform. There does not seem to exist a way to use relative sizes
inside a scrollable element that "ignore" the scrollbar width; e.g.
"width: 100%" refers to the "inside width" of the parent element with
the scrollbar width already subtracted, and there's no pure-CSS way to
say "set this element's width to the *outer width* of its parent element".