Fix Flow errors around NodeIndex and ClipboardEvent in the TreeView
I had previously typed the NodeIndex in the TreeView component as a
generic type bound as a number. I had assumed that this would be fine,
and I could treat everything as a structurally typed number. However,
the new errors from the previous commits coverage changes revealed that
it was treating the NodeIndex as an opaque type. This change removes the
generic type, and adds it as a normal number, aliased by NodeIndex.

Additionally, there was an error revealed around the ClipboardEvent.
I originally tried to add the ClipboardEvent type instead of inferring
it, but it broke the document.addEventListeners calls, as those
listeners did not like using the ClipboardEvent as an event handler.