Provide a generation value in state for invalidating panel layout sizing
This commit adds a panelLayoutGeneration value in the store that
increases monotonically on any change to the state that could
potentially invalidate the sizing of components in regards to
the panel layout.

These components include anything to do with the timeline height, and
with the sidebar openeing or being moved.

Resolves #1205
Resolves #736