Update the TracingMarkers to be used in the MarkerTable component
This commit re-organizes the store to use TracingMarkers to be used in
the marker table component. This makes it so that TracingMarkers start
to become the only markers that are actually used in the UI, since they
can be used more reasonably as both markers with and without durations,
plus the duplicated end markers will be successfully removed.