Update the StackGraph to use sample selection
This is in preparation of the ThreadActivityGraph from PR #1072. It
updates the stack graph to use sample selection.

The largest change from the original commits was to change the selection
state from a C enum-like value, with named variables, to strings. I felt
that this was more type-friendly for Flow, and shouldn't add any
overhead due to the nature of how JS creates strings.

Co-authored-by: Markus Stange <mstange@themasta.com>