Import Google Chrome profiles
This is not a complete solution, but a best effort to pull over the
samples an stacks from a Google Chrome performance profile. This work
should also include follow-ups to extract more information as needed.

This commit also includes some profile type refinement in order to deal
with importing foreign profiles.