Fixes flow errors + aligns chart
* Fix flow errors for the NetworkChart PR #1232

* Properly align the network panel rows

This commit changes the implementation from using the screen sizing, to
measuring the component with the WithSize higher order component. It
removed some CSS margins and padding that was getting in the way of the
layout as well.

Some of the methods were broken out, due to the VirtualListRow not
updating when the props were changing. I added a memoized function that
accepts the current props, and generates an array of the
NetworkChartRowProps that will properly trigger the VirtualListRow to
re-render. This update removed the React index values, as that is
handled by the VirtualListRow.

* Update snapshots for the network chart