Change the stack chart drawing strategy for small gaps
This PR changes the strategy for drawing the stack chart to draw small
boxes freely, but to leave 1 pixel gaps between boxes. If there are many
small boxes, it skips them until it finds either the next small box 1
pixel over, or the next box that is larger than 1 pixel.

In addition, it provides a small protection against calls to set
fillStyle which are slow due to the re-parsing of the CSS color.