Take null and undefined responsiveness into account for jank instance (Fixes #1472)
The chrome import script does not provide responsiveness values, and the
type was made optional. Previously the getJankInstances had coerced null
values into numeric. This patch corrects the usage to just skip null
values. It turns out that occasionally the sampler skips a
responsiveness value. Inspecting a profile with this behavior revealed
that it was best to just ignore those samples. In addition, the function
still works for completely null or undefined values.

Tests were also added to this patch explicitely test this behavior.