Remove the stack chart reducer, as it's not longer used
The stack chart reducer was added with the idea that we could change the
coloring strategies. This was done in the early days, and we've never
moved forward with this work. This commit removes this functionality as
the chart is just using the color categories. This is part of the work
to decrease the number of colors we have.