Upgrade Husky -- step 1 (PR #1825)
This is step 1 of the upgrade for our dependency Husky. Newer versions change how they pass the git parameters. So to make it easier to move between branches, so here is how we'll ugprade this dependency:

* step 1: change the scripts to support both names
* step 2: upgrade husky itself
* step 3: remove the layer compatibility

Between each of these steps we can wait a few days, maybe 2 weeks or a bit more.