Generalize the call tree to accept events that have stacks
This is a first step to generalize the call tree to handle multiple
different types of evented data. I believe the ultimate solution
is to generalize this to use an EventTable, however it's too early
at this time to abstract this logic. This current implementation
should be good enough.

See issue #2165 for more information about a perhaps better future
abstraction.