Add a mechanism to convert JS tracer data into threads
I made the decision to not do this conversion in a selector since
most of the UI makes an assumption about a stable thread list. This
is a bit riskier because the conversion gets baked into the profile
itself. I feel like this makes dealing with JS tracer data in this
way easier for the implementation, and the risk is low because this
is an extremely experimental feature.