Ignore network markers while getting non-global tab filtered markers to see if thread is empty
Previously we didn't have innerWindowIDs inside the network markers but
now we have that field in them. While we want to get the non-global
markers for a thread, we shouldn't get network markers because they can
be in the main thread even though, that network is being requested from
another thread. That else clause is only being executed while we are
determining if the thread is empty or not.