Only create the SymbolStore if IndexedDB exists
The test suite was spitting our random promise failure errors on code
whenever the SymbolStore was being accessed without mocking out
IndexedDB. This patch changes the behavior where symbolication is
only attempted if the IndexedDB is actually present. This way
tests don't need random mocking of the IndexedDB value when
it's not really needed. This was happening anytime a Gecko Profile
was being used in a test.