Fix the performance issue on flow server initialization
Before this patch flow server was taking ~100 seconds to initialize.
After this patch, it takes ~10 seconds to initialize on my machine.
This is most likely because we cannot overload the methods inside the
extended classes(EventTarget in that case). Because it looks like this
performance issue appeared after the `addEventListener` and
`removeEventListener` signature change for webchannels.

This is the EventTarget type in flow:
https://github.com/facebook/flow/blob/2e1987fd9a5e7fa21859dfa6de621c4f4a2162ae/lib/dom.js#L199-L274
`attachEvent` and `detachEvent` methods are non-standards and others are
deprecated already. Only other method that we need is `dispatchEvent`.