Use Profile trace events to treat ProfileChunk data properly.
The ProfileChunk sampling data was assigned to the wrong thread and shifted in time,
misaligned with the profile markers.
To get the correct time alignment and thread index, use the Profile event:
Each Profile event is associated with a number of ProfileChunk events via a shared object id.
It's the Profile event that has the correct startTime and the correct tid.
The 'ts' timestamps on the ProfileChunk events should not be used to compute sample times.