Consolidate symbolication updates.
This change prepares the code for a future change that will handle
function splitting during re-symbolication.
This change also adjusts the canonical func's address to point to the
function symbol. (Though I realize now that this is not a good change
to make in this commit, both because it's implemented in a way that
breaks funcTable immutability, and because it makes re-symbolication with
the current func-based address gathering impossible. But all of that will
be fixed in the next commit.)