Pass array buffers to unserializeProfileOfArbitraryFormat.
This allows unserializeProfileOfArbitraryFormat to handle more input formats.
In order to parse binary formats such as the ART trace format, we cannot just
parse files as text, we need to pass the actual binary data along.

This change also reports more source errors to the console, so that their
stacks aren't lost.

And this change makes utf-8 decoding errors cause exceptions. In the post, these
would be swallowed silently, because TextDecoder's default is to be non-fatal.