Use getState() to compute oldFuncToNewFuncMap at a better time.
The previous commits were doing symbol substitution in two halves:
The first half was happening in finishSymbolicationForLib,
which couldn't touch the thread and couldn't create new funcs.
But finishSymbolicationForLib needed to create an oldFuncToNewFuncMap
that could be stuffed into the dispatched action.
And then the second half was happening in applySymbolicationStep.
This was called from the profile reducer and could make the actual
changes to the thread tables.

Now, all of this happens in applySymbolicationStep, and applySymbolicationStep
is no longer called from the reducer. Instead, it's called from the
'coalescedSymbolicationStep' thunk action. It computes the new thread and stuffs
it into an action, and then the profile reducer just takes the finished
threads list and places it in the profile.
This allows applySymbolicationStep to create the oldFuncToNewFuncMap,
which means it can now take into account any new funcs that needed to
be created on the way.

This commit also renames some things around "coalesced symbolication step"
handling, which used to be called "coalesced functions update".