Improve distinction of MemoryOffset and Address.
Clarify that Address values are always relative to some library.
Only use MemoryOffset for absolute addresses.
I didn't add a LibRelativeOffset type because it's a mouthful and it's the
type we'd use almost everywhere. "AddressÆ is nicer to type and to read.

Also improve Lib documentation.

Some comments are a bit vague about how many mappings of a library there can be.
That's mostly because it's something I'm unsure about myself.
I think there can be many mappings of different parts of one library into one process,
and each mapping can have different read/write/execute permissions. And I think the
singular range that we report for a library from shared-libraries-linux.cc is the
union of the ranges of all the mappings. The things I'm unsure about is:
 - Can there be gaps? Do we care?
 - We rely on the fact that there's a global "baseAddress" of each library
   in the profiled process's address space. Is that always the case? Or could
   different mappings be mapped in such a way that the base address would be
   different between them?
We haven't run into any problems from this yet so I haven't looked into it more closely.