Remove the setIsInitialSelectedPane prop from resource tracks
That was doing nothing. That was something we needed for full view
global tracks because we were selecting the parent track if local track
is selected but this doesn't apply to active tab anymore. We don't need
to select the parent since there is only one parent. We can just remove
this code piece.