Add support for displaying extra IPC markers; see BMO bug 1613496
Firefox now creates up to 5 profiler markers for each IPC message. In addition
to the previously-existing "start" and "end" markers, we now have ones for
"starting to send bytes", "finished sending bytes", and "finished receiving
bytes". These are all correlated together to provide time deltas between each
phase of the IPC message.

More documentation for how to use this feature is available in
<docs-user/ipc-messages.md>.