Fix the scrollbar problem on active tab
Scrollbar was not showing up properly on active tab due to this css
variable. That's because we are using this value inside the `calc`
function and it expects this value to be a pixel. For example, this
is the one use case of this variable:
`margin-left: calc(var(--thread-label-column-width) * -1);`
If we put an integer to this variable instead, calc result will be
integer and this margin propery will fail to compute. So we *must* use a
pixel for this variable even though it's `0`.