Refine the marker schema for the tooltips
This commit has a few things going on, but I was finding it difficult to
split everything out. It adds a new mechanism for computing a tooltip
label, with printf like functionality. It was briefly suggested in the
design doc under a comment, and I was initially reluctant to take it on
due to the complexity. However, I was struggling with implementing
things a certain way, and I found it useful.

The selector approach for the label maker may seem like a bit overkill,
but I'm planning on using it in hot locations in the future.

In addition, this commit centralizes how the marker schema name is
determined. There are a few a ad-hoc rules around how marker schema gets
applied, such as with tracing markers. There should be TODOs in place
with references to issues.

Finally these changes are actually applied to the marker Tooltip
component.