Update and remove a few babel dependencies that were stuck to the old v6 (PR #2778)
This removes the specific configuration for dynamic imports, because
all node, babel and webpack seem to support it natively.

This also removes the @babel/runtime direct dependency which seems to
be a left-over. Indeed it's supposed to be used in tandem with the
transform which wasn't configured in our project. This is a tool to be
able to use newer JavaScript APIs, but because we take care to use only
what's in Firefox and Chrome we shouldn't need it anyway.