Remove the marker title
This leaves the title in for Jank and IPC markers, which will be handled
in an upcoming commit. This is also probably a partial commit, as the
title will fully be replaced by the new marker schema labels. This
commit is mostly the mechanical action of removing the title field.