Handle out of range scenarios in bisectionLeft and bisectionRight
For the bisection functions, low and high values should lie within the range of the array. This commit adds a check to the bisection functions to throw TypeError in case the low or high values passed are negative or greater than the array length.
This commit also adds relevant testcases to bisect.test.js to check these conditions.