Remove interval: "start" | "end" from tracing payload
This property was removed awhile back, but our types still reference it.
This patch removes it. Most of the changes are reflected in the tests.
The only place where it's touching production code is the chrome
importer. However, that chrome importer is correctly using the new
tracing markers, and does not rely on this interval property. No
upgrader is needed as this is only bringing our types up to date with the
current reality.

Current tracing payload:

https://searchfox.org/mozilla-central/rev/168c45a7acc44e9904cfd4eebcb9eb080e05699c/mozglue/baseprofiler/public/BaseProfilerMarkers.h#157-161