Improve error reporting.
This catches any exceptions during ART format parsing and throws an error
with a message that mentions the fact that we were trying to parse an ART
trace. The original exception is logged to the console and its message is
included in the re-wrapped error.

What ends up being exposed to the user is a concatenation of a whole
number of error messages. For example:

```
Error: Unserializing the profile failed: Error: Could not parse the
profile array buffer as an ART trace: Error: This code only knows how to
parse versions 1 to 3, got version 5.
```