Add tests for the split samples rendering in the activity graph with tooltip
Even though technically we are testing the activity graph here, we are
checking this in the tooltip because this is the easiest way to check the
activity graph. Otherwise the current tests for activity graph is mostly
about snapshots and they are not really expressive enough when it comes
to seeing what changed.