Adjust return addresses.
For return addresses, look up addresses minus one byte, to fall into the
previous instruction. This is necessary to get accurate line information
and accurate inline stacks.
The stack leaf (instruction pointer) must not be adjusted.
As a result, this is a rather invasive process because the stack table
needs major surgery to account for new stacks.