Move function address from funcTable into a new nativeSymbols table.
This is a change to the processed format and requires a new version.

No user-observable behavior differences are expected. The only thing
that could happen is that functions with identical names but at
different symbol addresses now get merged into the same call node when
they would have been separate before, but you usually never see such
identically-named functions.

Here's what the patch does.

Brief recap: For native stack frames, each "frame" in a thread's
frameTable corresponds to one "address": The address of the instruction
that was being executed at the time of sampling. During symbolication,
we usually find out that many of those separate frames actually belong
to a much smaller set of functions. So we group frames by their
function, and compute a call tree based on this grouping.
At the moment, we create functions based on their function symbol
address: Each unique symbol address is treated as an individual function,
with the symbol name being used as the function name.

Creating functions based on symbols works fine for now, but it won't work
if we want to support inline call stacks, because inline callstacks can
contribute functions which don't have their own symbol. For example,
let's say I have a frame at address 0x1005, and the closest symbol is the
symbol "OuterFunction()" at address 0x1000. With inline callstacks, the
symbolication API can give me a result such as:

> The inline call stack at address 0x1005 is:
> InnermostFunction()
> MiddleWrapperFunction()
> OuterFunction()

So now we need to create entries in the funcTable for
"InnermostFunction()" and "MiddleWrapperFunction()" even though those
functions don't have their own symbol in the library's symbol table.
And if InnermostFunction() is also inlined somewhere else, into another
outer function at a different symbol address, then those two
"InnermostFunction()" invocations should still share the same func in
spite of the different addresses. In other words, it would be better to
create functions based on their name.

So this patch removes the funcTable's address field, and instead creates
funcs based on function names. Now all frames that are resolved to the
same function name will share the same func, even if two functions with
identical names have different symbol addresses.

And to keep information about the symbol address in the profile, a new
table is added to the thread, called thread.nativeSymbols. The symbol
address information isn't really used for the UI at the moment, but it
will be needed once we add an assembly view.

The patch is rather large, unfortunately. Most of it is boilerplate which
comes with every format change. Only the changes in symbolication.js are
actually interesting. Symbolication still has the symbol address based
grouping (which now operates on nativeSymbols instead of on funcs), and
it has a new section where we create funcs based on the function name.

The easiest way to review it is probably to start by looking at the type
definitions in src/types/profile.js and at the large comment at the top
of symbolication.js.