Don't send symbolication requests with more than 10 jobs.
The new symbolication server has a limit of 10 jobs per request.
This limit applies regardless of the number of stacks per job.
In the past, we wouldn't run into this limit because we requested
one job with many stacks (one stack for each library). Now that we
request one *job* per library (with one stack each), we run into
the limit.
We could probably ask to have the limit changed, but symbolicating
10 libraries per request seems like a reasonable chunking size to
have anyway.