Use :focus-visible in place of :focus.
This eliminates extraneous focus rings on macOS and Windows. For example, clicking
buttons, checkboxes and thread tracks will no longer show a focus ring on those
platforms. Tabbing to these controls will still show a focus ring. And clicking
text boxes will show a focus ring, too.
On macOS, this is basically a regression fix for a recent change in Firefox behavior:
https://bugzilla.mozilla.org/show_bug.cgi?id=1614658 (shipped in Firefox 88) made
it so that focus on macOS behaves just like focus on Windows: Clicking always focuses
but usually doesn't show the focus ring. Whereas in the past, clicking wouldn't even
have focused; focus was kept in the last place that the user had tabbed to. Anyway,
starting with Firefox 88, the profiler started showing focus rings in many more
places than before, when run on macOS. And this change fixes it.

:focus-visible is supported in Firefox 85 and above and Chrome/Edge 86 and above.
It is not supported in Safari yet. But in Safari, any controls that get a native
focus ring will still show the native focus ring when appropriate, because our
styles do not unset the native focus ring (unless :focus-visible is supported).
So the fallback behavior is actually surprisingly acceptable.

To get perfect backwards compatibility, we'd need to do a 'set, unset, set' dance
with three different CSS selectors for every selector. That's too much work.

There's one :focus that we keep: the one for the treeview which uses tabindex='0'.