Use a more compact URL encoding for track indexes.
This massively cuts down the URL length for profiles with a large number of tracks;
rather than having "0-1-2-3-4-5" in the URL, we just have "0w5".

This is achieved by adding a special case for "consecutive ranges" to
uintarray-encoding (exploiting "leading zero" digits of the VLQ) and
switching the track index URL state properties to use that encoding.

This commit also improves the documentation for the UintArray string
encoding module and adds a small test for it.