Run the correct set of URL version upgraders.
_upgraders[4] should only be called when upgrading a URL *to* version 4.
But due to a bug, it was also called when upgrading *from* version 4.
It seems that this was not a problem in the past because all upgraders
were idempotent. But the upgrader for version 5 is not idempotent; running
it on a URL of version 5 discards the range filter.

I noticed this bug when trying to add a version 6; all old URLs lost their range filter.