Properly store the combined thread transform in the URL.
Fixes #3103.

When multiple threads are selected, we display a merged thread. When transforms are
added in this view, they affect the merged thread. The redux state separately stores
the transform stack for each individual thread, and also any transform stack for combined
threads, in a TransformStacksPerThread object. For example, if I have three threads, the
object might look like this:

{
  '0': transforms for thread 0,
  '1': transforms for thread 1,
  '0,2': transforms for the combined thread when threads 0 and 2 are selected
}

The URL does not store this full object, it only stores the transform stack for the
current view, i.e. the transform stack for the selected thread.

This wasn't working properly when multiple threads were selected: In the above example,
if threads 0 and 2 were selected, the URL was storing the transforms for the individual
threads 0 and 2. However, those transforms do not affect the merged thread, so they're not
relevant for the current view and should not be put into the URL.
Instead, the URL should store the transform for the merged thread '0,2'. This patch fixes
that. Doing this is simpler, too.