Make sure old mousemove/mouseup handlers are cleaned up.
This prevents us from getting stuck in drag-selection mode if, for some reason,
the mouseup event never fires.

One way this can happen is if there's a tab switch in the middle of the drag
gesture. To reproduce, start a drag selection and then press ctrl+tab and then
let go of the mouse button. Switch back to the profiler tab, drag-select again,
and see if drag selection persists after the mouseup.