Switch many uses of mouseup to use click instead.
Fixes #3451.

Click has the advantage that it only fires if the corresponding mousedown
also occured over this element. That means that it does not fire if the
mouse was dragged in from outside this element.
Using click instead of mouseup avoids unintentionally switching the
selected thread when panning a chart, for example.

I'm not sure why many of these used mouseup to begin with.
It's possible that I had something in mind when I created the first
components with mouseup, and that whatever I was thinking no longer
applies. Most of the components that are using mouseup today are
probably using it because other existing components were using it.

But it is possible that we will find certain interactions which are
broken by this patch. I've done some testing but not a huge amount, and
would appreciate more testing.