Use uintarray encoding for the selected tracks.
It was using 4,6,8 before, but the commas don't look pretty when they're URL-encoded.
This syntax is also more compact when a range of threads is selected, for example
when all stylo worker threads are selected.

The upgrader piggy-backs on the same new URL version that was added in PR #3419, because
that PR hasn't been deployed yet so there shouldn't be any URLs in circulation with it.