Enforce max-width: 600px for all tooltips.
This patch removes the CSS variable --tooltip-detail-max-width. The
maximum width is no longer configurable on a per-tooltip basis or
overridable to different values for different parts of a tooltip.
Instead, all tooltips are constrained to the same width, 600px.
These freedoms weren't used anywhere; only flame graph tooltips were
unconstrained, probably by mistake.

With this patch, the max-width constraint is no longer applied to
individual "details" parts, such as a backtrace or a URL. Instead, the
max-width is applied on the root tooltip element, and the contents are
constrained via grid / flexbox, sometimes in combination with min-width:0.
The old implementation sometimes caused broken layouts where, for example,
the overall width of a network tooltip was set to 600px, and then the
backtrace inside the network tooltip had a max-width:600px which only
affected the right column of a two-column tooltip layout. This is now
fixed.

The patch also includes some changes to enforce consistent line breaking
of URLs in various tooltips.

The most obvious result from this patch is that tooltips in the Flame
Graph tab are now constrained. They weren't constrained before.