Update remote profiling instructions.
This fixes a broken link to the 32 bit version of GeckoView-example; the link was
pointing to a two month old build, because the route to the latest build was changed
in https://phabricator.services.mozilla.com/D114369 .
Thanks to bhearsum for finding the change and the new route!

I'm also modernizing some other content in this file.